refactor: server decoupling
