--- .~lock.Pokemon Go.xlsx#
@@ -1,1 +1,0 @@
-,EL/ana.mendes,WKS-DEV207-DGM,16.04.2023 19:03,file:///C:/Users/ana.mendes/AppData/Roaming/LibreOffice/4;
